# docs/nest-store.swagger.json
{"info":{"title":"nest-store","description":"","version":"1.0.0"},"tags":[{"name":"user"},{"name":"gift"}],"paths":{"/user":{"get":{"summary":"list user","deprecated":false,"description":"","tags":["user"],"parameters":[],"responses":{"200":{"description":"Success","schema":{"type":"object","properties":{}}}},"security":[],"produces":["application/json"]},"post":{"summary":"add user","deprecated":false,"description":"","tags":["user"],"parameters":[{"name":"body","in":"body","schema":{"type":"object","properties":{"name":{"type":"string","faker":"name.fullName"},"email":{"type":"string","faker":"internet.email"},"password":{"type":"string","faker":"internet.password"}},"required":["name","email","password"]}}],"responses":{"200":{"description":"Success","schema":{"type":"object","properties":{}}}},"security":[],"consumes":["application/json"],"produces":["application/json"]}},"/auth/login":{"post":{"summary":"login","deprecated":false,"description":"","tags":[],"parameters":[{"name":"body","in":"body","schema":{"type":"object","properties":{}}}],"responses":{"200":{"description":"Success","schema":{"type":"object","properties":{}}}},"security":[],"consumes":["application/json"],"produces":["application/json"]}},"/gift":{"get":{"summary":"list-gift","deprecated":false,"description":"","tags":["gift"],"parameters":[{"name":"page","in":"query","description":"","required":false,"type":"string","x-example":"1"},{"name":"limit","in":"query","description":"","required":false,"type":"string","x-example":"10"},{"name":"sortBy","in":"query","description":"","required":false,"type":"string","x-example":"createdAt"},{"name":"order","in":"query","description":"","required":false,"type":"string","x-example":"DESC"}],"responses":{"200":{"description":"Success","schema":{"type":"object","properties":{}}}},"security":[{"bearer":[]}],"produces":["application/json"]},"post":{"summary":"create gift","deprecated":false,"description":"","tags":["gift"],"parameters":[{"name":"body","in":"body","schema":{"type":"object","properties":{}}}],"responses":{"200":{"description":"Success","schema":{"type":"object","properties":{}}}},"security":[{"bearer":[]}],"consumes":["application/json"],"produces":["application/json"]}},"/gift/{id}":{"get":{"summary":"detail gift","deprecated":false,"description":"","tags":["gift"],"parameters":[{"name":"id","in":"path","description":"","required":true,"type":"string","x-example":"22d77e0f-eaec-4081-bcd7-bca67aa253c6"}],"responses":{"200":{"description":"Success","schema":{"type":"object","properties":{}}}},"security":[{"bearer":[]}],"produces":["application/json"]},"patch":{"summary":"update stock","deprecated":false,"description":"","tags":["gift"],"parameters":[{"name":"id","in":"path","description":"","required":true,"type":"string","x-example":"22d77e0f-eaec-4081-bcd7-bca67aa253c6"},{"name":"body","in":"body","schema":{"type":"object","properties":{}}}],"responses":{"200":{"description":"Success","schema":{"type":"object","properties":{}}}},"security":[{"bearer":[]}],"consumes":["application/json"],"produces":["application/json"]},"put":{"summary":"update gift","deprecated":false,"description":"","tags":["gift"],"parameters":[{"name":"id","in":"path","description":"","required":true,"type":"string","x-example":"22d77e0f-eaec-4081-bcd7-bca67aa253c6"},{"name":"body","in":"body","schema":{"type":"object","properties":{}}}],"responses":{"200":{"description":"Success","schema":{"type":"object","properties":{}}}},"security":[{"bearer":[]}],"consumes":["application/json"],"produces":["application/json"]},"delete":{"summary":"delete gift","deprecated":false,"description":"","tags":["gift"],"parameters":[{"name":"id","in":"path","description":"","required":true,"type":"string","x-example":"22d77e0f-eaec-4081-bcd7-bca67aa253c6"}],"responses":{"200":{"description":"Success","schema":{"type":"object","properties":{}}}},"security":[{"bearer":[]}],"produces":["application/json"]}},"/user/{id}":{"get":{"summary":"detail-user","deprecated":false,"description":"","tags":["user"],"parameters":[{"name":"id","in":"path","description":"","required":true,"type":"string"}],"responses":{"200":{"description":"Success","schema":{"type":"object","properties":{}}}},"security":[],"produces":["application/json"]}},"/gift/{id}/redeem":{"post":{"summary":"redeem","deprecated":false,"description":"","tags":["gift"],"parameters":[{"name":"id","in":"path","description":"","required":true,"type":"string","x-example":"8ae4e68e-2797-4816-beae-d76a7ab2ad5e"},{"name":"body","in":"body","schema":{"type":"object","properties":{}}}],"responses":{"200":{"description":"Success","schema":{"type":"object","properties":{}}}},"security":[{"bearer":[]}],"consumes":["application/json"],"produces":["application/json"]}},"/gift/{id}/rating":{"post":{"summary":"add rating","deprecated":false,"description":"","tags":["gift"],"parameters":[{"name":"id","in":"path","description":"","required":true,"type":"string","x-example":"8ae4e68e-2797-4816-beae-d76a7ab2ad5e"},{"name":"body","in":"body","schema":{"type":"object","properties":{}}}],"responses":{"200":{"description":"Success","schema":{"type":"object","properties":{}}}},"security":[{"bearer":[]}],"consumes":["application/json"],"produces":["application/json"]}}},"swagger":"2.0","host":"localhost:3001","schemes":["http"],"basePath":"/","definitions":{"CreateUserDto":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"}},"required":["name","email","password"]},"CreateUserDto1":{"type":"object","properties":{"name":{"type":"string","faker":"name.fullName"},"email":{"type":"string","faker":"internet.email"},"password":{"type":"string","faker":"internet.password"}},"required":["name","email","password"]}},"securityDefinitions":{"bearer":{"type":"apiKey","name":"Authorization","in":"header"}},"x-components":{}}